Added a separate UI area for user position, next step is to add a button to start and stop the polling service.
