+++ app/src/main/res/layout/fragment_main.xml
@@ -6,6 +6,8 @@
     android:layout_height="match_parent"
     tools:context=".MainFragment">
 
+    <include layout="@layout/user_pos"
+        android:id="@+id/userPos"/>
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerView"
@@ -14,8 +16,9 @@
         android:clickable="true"
         android:focusable="true"
         android:focusableInTouchMode="true"
+        app:layout_constraintTop_toBottomOf="@id/userPos"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+         />
 </androidx.constraintlayout.widget.ConstraintLayout>
