+++ app/src/main/res/values/strings.xml
@@ -12,4 +12,11 @@
     <string name="password_hint">Enter your password</string>
     <string name="settings_title">Settings</string>
     <string name="save_btn_label">Save settings</string>
+    <string name="last_updated_at">Last updated at:</string>
+    <string name="last_updated_date">on:</string>
+    <string name="lat">Lat:</string>
+    <string name="lon">Lon:</string>
+    <string name="alt">Alt:</string>
+    <string name="acc">Acc:</string>
+    <string name="user_position">You Are Here!</string>
 </resources>
